sensor/utils.py
import os
import sys
import yaml
import dill
import numpy as np
import pandas as pd
from sensor.exception import SensorException
from sensor.logger import logging
from sensor.config import mongo_client

def get_collection_as_dataframe(database_name:str,collection_name:str)->pd.DataFrame:
    '''
    Description : This function returns collections as dataframe 
    ===================================================================
    Params : 
    Database name : database_name
    Collection_name : collection_name 
    ===================================================================
    return Pandas dataframe of the collection  
    '''
    try:
        logging.info(f"Reading data from database_name {database_name} and collection {collection_name}")
        print("reading data from mongo db")
        df=pd.DataFrame(list(mongo_client[database_name][collection_name].find()))
        print("Data frame created ")
        logging.info(f"Columns present are {df.columns}")
        if "_id"in df.columns:
            logging.info("Removing id column from dataframe ")
            df.drop('_id',axis=1,inplace=True)
        logging.info(f"The shape of the data is {df.shape}")
        return df
    except Exception as e:
        raise SensorException(e,sys)
    
def write_yaml_file(file_path,data:dict):
    try:
        file_dir = os.path.dirname(file_path)
        os.makedirs(file_dir,exist_ok=True)
        with open(file_path,"w") as file_writer:
            yaml.dump(data,file_writer)
    except Exception as e:
        raise SensorException(e, sys)

def convert_columns_float(df:pd.DataFrame,exclude_columns:list)->pd.DataFrame:
    try:
        for column in df.columns:
            if column not in exclude_columns:
                df[column]=df[column].astype('float')
        return df
    except Exception as e:
        raise e
    
def save_object(file_path: str, obj: object) -> None:
    try:
        logging.info("Entered the save_object method of utils")
        os.makedirs(os.path.dirname(file_path), exist_ok=True)
        with open(file_path, "wb") as file_obj:
            dill.dump(obj, file_obj)
        logging.info("Exited the save_object method of utils")
    except Exception as e:
        raise SensorException(e, sys) from e


def load_object(file_path: str, ) -> object:
    try:
        if not os.path.exists(file_path):
            raise Exception(f"The file: {file_path} is not exists")
        with open(file_path, "rb") as file_obj:
            return dill.load(file_obj)
    except Exception as e:
        raise SensorException(e, sys) from e

def save_numpy_array_data(file_path: str, array: np.array):
    """
    Save numpy array data to file
    file_path: str location of file to save
    array: np.array data to save
    """
    try:
        dir_path = os.path.dirname(file_path)
        os.makedirs(dir_path, exist_ok=True)
        with open(file_path, "wb") as file_obj:
            np.save(file_obj, array)
    except Exception as e:
        raise SensorException(e, sys) from e
    


def load_numpy_array_data(file_path: str) -> np.array:
    """
    load numpy array data from file
    file_path: str location of file to load
    return: np.array data loaded
    """
    try:
        with open(file_path, "rb") as file_obj:
            return np.load(file_obj)
    except Exception as e:
        raise SensorException(e, sys) from e